Select Multiple Terrain Tiles In Sequence
Added the support for selecting Terrain tiles in sequence. Still need
more work on what happens when units are selected and then a click
occurs. Will need to add more states like 'moving' 'attacking' or
something along those lines. Maybe just generating activities is enough.
Probably dependent at least in part on UI concerns, but I think in
general the controller needs to know what the unit is trying to
accomplish so that it can tell the model what to do with the actions.