Moving On Up
Implemented a fair bit of movement oriented code. Reached the point
where I need to implement a bit of pathfinding, but I'm still shaky on
how exactly to go about doing that. At this point, I'll continue
researching A* and get into a quick chat with Lennon about it when he
has some time.